backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
passlib[bcrypt]
pyjwt
python-multipart
cryptography
pymongo
